add Resource types
